requirements.txt
chromadb==1.2.2
langchain-chroma==1.0.0
langchain-community==0.4.1
langchain-google-genai==2.1.0
langchain-text-splitters==1.0.0
pypdf==6.1.3
python-dotenv==1.2.1
